fix merge conflicts issues
