+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/ListType.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.pip.publication.services.models.external;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.converters.CivilAndFamilyDailyCauseListConverter;
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.converters.CivilDailyCauseListConverter;
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.converters.Converter;
