+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/PdfCreationService.java
@@ -10,6 +10,7 @@ import org.thymeleaf.context.Context;
 import org.thymeleaf.spring5.SpringTemplateEngine;
 import uk.gov.hmcts.reform.pip.publication.services.config.ThymeleafConfiguration;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.Artefact;
+import uk.gov.hmcts.reform.pip.publication.services.models.external.Location;
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.converters.Converter;
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.helpers.Helpers;
 
@@ -43,10 +44,14 @@ public class PdfCreationService {
     public String jsonToHtml(UUID inputPayloadUuid) throws IOException {
         String rawJson = dataManagementService.getArtefactJsonBlob(inputPayloadUuid);
         Artefact artefact = dataManagementService.getArtefact(inputPayloadUuid);
+        Location location = dataManagementService.getLocation(artefact.getLocationId());
+
         String htmlFile;
         Map<String, String> metadataMap =
             Map.of("contentDate", Helpers.formatLocalDateTimeToBst(artefact.getContentDate()),
-                   "provenance", artefact.getProvenance(), "location", artefact.getLocationId());
+                   "locationName", location.getName(),
+                   "provenance", artefact.getProvenance(),
+                   "location", artefact.getLocationId());
 
         JsonNode topLevelNode = new ObjectMapper().readTree(rawJson);
