+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/pdf/helpers/PartyRoleMapper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pip.publication.services.service.helpers;
+package uk.gov.hmcts.reform.pip.publication.services.service.pdf.helpers;
 
 import java.util.List;
 import java.util.Map;
