+++ src/main/resources/templates/familyCauseList.html
@@ -34,8 +34,6 @@
     .govuk-heading-l {
       color: #0b0c0c;
       font-family: "GDS Transport";
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
       font-weight: 700;
       font-size: 24px;
       font-size: 1.5rem;
@@ -47,8 +45,6 @@
     .govuk-body {
       color: #0b0c0c;
       font-family: "GDS Transport";
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
       font-weight: 400;
       font-size: 16px;
       font-size: 1rem;
@@ -61,8 +57,6 @@
     }
     .govuk-details {
       font-family: "GDS Transport", arial, sans-serif;
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
       font-weight: 400;
       font-size: 16px;
       font-size: 1rem;
@@ -98,7 +92,7 @@
       top: 2em;
     }
     .site-address {
-      margin-block-end: 0.5em;
+      margin-top: 0.5em;
     }
     .govuk-accordion__section-heading {
       margin-top: 0;
@@ -106,8 +100,6 @@
       padding-top: 15px;
       padding-bottom: 15px;
       font-family: "GDS Transport", arial, sans-serif;
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
       font-weight: 700;
       font-size: 18px;
       font-size: 1.125rem;
@@ -120,8 +112,6 @@
     }
     .govuk-table {
       font-family: "GDS Transport", arial, sans-serif;
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
       font-weight: 400;
       font-size: 16px;
       font-size: 1rem;
@@ -214,8 +204,8 @@
                 </tr>
                 </thead>
                 <tbody class="govuk-table__body">
-                  <span th:each="sitting : ${session.get('sittings')}">
-                    <span th:each="hearing : ${sitting.get('hearing')}">
+                  <th:block th:each="sitting : ${session.get('sittings')}">
+                    <th:block th:each="hearing : ${sitting.get('hearing')}">
                       <tr class="govuk-table__row" th:each="case : ${hearing.get('case')}">
                         <td class="govuk-table__cell no-wrap" th:text="${sitting.get('time').asText()}" />
                         <td class="govuk-table__cell no-wrap" th:text="${case.get('caseNumber').asText()}" />
@@ -227,8 +217,8 @@
                         <td class="govuk-table__cell" th:text="${hearing.get('applicant').asText()}" />
                         <td class="govuk-table__cell" th:text="${hearing.get('respondent').asText()}" />
                       </tr>
-                    </span>
-                  </span>
+                    </th:block>
+                  </th:block>
                 </tbody>
               </table>
             </div>
