+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/PdfCreationServiceTest.java
@@ -12,6 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.Artefact;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.ListType;
+import uk.gov.hmcts.reform.pip.publication.services.models.external.Location;
 
 import java.io.IOException;
 import java.io.StringWriter;
@@ -75,10 +76,14 @@ class PdfCreationServiceTest {
         artefact.setLocationId("1");
         artefact.setProvenance("france");
         artefact.setListType(ListType.MAGS_STANDARD_LIST);
+        Location location = new Location();
+        location.setName("location");
+
         UUID uuid = UUID.randomUUID();
         String inputJson = "{\"document\":{\"value1\":\"x\",\"value2\":\"hiddenTestString\"}}";
         when(dataManagementService.getArtefactJsonBlob(uuid)).thenReturn(inputJson);
         when(dataManagementService.getArtefact(uuid)).thenReturn(artefact);
+        when(dataManagementService.getLocation("1")).thenReturn(location);
 
         byte[] outputPdf = pdfCreationService.generatePdfFromHtml(pdfCreationService.jsonToHtml(uuid));
         try (PDDocument doc = PDDocument.load(outputPdf)) {
